+++ src/layouts/DocLayout/GlobalStyle.ts
@@ -1,8 +1,13 @@
 import { createGlobalStyle } from 'antd-style';
 
 const GlobalStyle = createGlobalStyle`
+  html, body {
+    background: ${({ theme }) => theme.colorBgLayout};
+  }
+
   #root {
     min-height: 100vh;
+    background: ${({ theme }) => theme.colorBgLayout};
   }
 
   #nprogress {